moving soql editor event into own file
