+++ packages/soql-builder-ui/src/modules/querybuilder/services/message/soqlEditorEvent.ts
@@ -1,0 +1,6 @@
+import { ToolingModelJson } from '../toolingModelService';
+
+export interface SoqlEditorEvent {
+  type: string;
+  message: string | ToolingModelJson;
+}
